Secure Files
Removes hardcoded passwords and simplifies functions for web scraper
