+++ .gitignore
@@ -45,3 +45,5 @@ $RECYCLE.BIN/
 Network Trash Folder
 Temporary Items
 .apdisk
+basic_beautiful_soup/secure_scraper_setup.zip
+stat_projections/neural_networks/nba_neural_network_projector.py
